# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Golang work.",
        "title": "Go-NOTE",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "zhangxin",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "http://127.0.0.1:8880/swagger/index.html",
    "basePath": "/api/v1",
    "paths": {
        "/login": {
            "post": {
                "description": "登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "登录并返回token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{object}\tModel.ResponseMessage"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "注册用户名和密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "邮箱",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/todo": {
            "get": {
                "responses": {
                    "200": {
                        "description": "{object} ResponseMessage"
                    },
                    "500": {
                        "description": "{object} ResponseMessage"
                    }
                }
            },
            "put": {
                "description": "更新记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "更新所有未完成或已完成记录的完成状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "完成状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{object} ResponseMessage"
                    },
                    "500": {
                        "description": "{object} ResponseMessage"
                    }
                }
            },
            "post": {
                "description": "添加记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "添加一条记录",
                "parameters": [
                    {
                        "description": "标题",
                        "name": "title",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "内容",
                        "name": "content",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{object} ResponseMessage"
                    },
                    "500": {
                        "description": "{object} ResponseMessage"
                    }
                }
            },
            "delete": {
                "description": "删除记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "删除所有/已完成/未完成记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "完成状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{object} ResponseMessage"
                    },
                    "500": {
                        "description": "{object} ResponseMessage"
                    }
                }
            }
        },
        "/todo/{title}": {
            "get": {
                "description": "查询关键词为title的记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "查询一条记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "标题",
                        "name": "title",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{object} ResponseMessage"
                    },
                    "500": {
                        "description": "{object} ResponseMessage"
                    }
                }
            },
            "put": {
                "description": "更新一条标题为title的记录，状态改为status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "更新一条记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "修改状态",
                        "name": "status",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "被更新记录的标题",
                        "name": "title",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{object} ResponseMessage"
                    },
                    "500": {
                        "description": "{object} ResponseMessage"
                    }
                }
            },
            "delete": {
                "description": "删除标题为title的记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "删除一条记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "被删除记录的标题",
                        "name": "title",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{object} ResponseMessage"
                    },
                    "500": {
                        "description": "{object} ResponseMessage"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Data": {
            "type": "object",
            "properties": {
                "item": {
                    "description": "数据信息",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Note"
                    }
                },
                "total": {
                    "description": "数据数目",
                    "type": "integer"
                }
            }
        },
        "models.Note": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "记录内容",
                    "type": "string"
                },
                "create_time": {
                    "type": "integer"
                },
                "donestatus": {
                    "description": "完成状态",
                    "type": "integer"
                },
                "end_time": {
                    "description": "截止时间",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "integer"
                },
                "title": {
                    "description": "记录题目",
                    "type": "string"
                }
            }
        },
        "models.ResponseMessage": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "数据信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Data"
                        }
                    ]
                },
                "err": {
                    "description": "错误信息",
                    "type": "string"
                },
                "msg": {
                    "description": "响应信息",
                    "type": "string"
                },
                "status": {
                    "description": "状态码",
                    "type": "integer"
                }
            }
        }
    }
}